seperate file for two algorithms
